requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
python-multipart
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
selenium
webdriver-manager
